requirements.txt
transformers==4.31.0
datasets==2.12.0
rouge_score==0.1.2
streamlit==1.24.0
pyyaml==6.0
torch==2.0.1
evaluate
accelerate==0.20.1